requirements.txt
pandas
pillow
tqdm
matplotlib
tensorboard
scikit-learn
torchmetrics
onnx
onnxscript
onnxruntime
